Move create_complete_tx_with_prefix
