+++ bin/citrea/tests/bitcoin/utils.rs
@@ -1,12 +1,8 @@
 use std::path::PathBuf;
-use std::str::FromStr;
 use std::sync::Arc;
 
-use bitcoin::secp256k1::SecretKey;
-use bitcoin_da::helpers::builders::body_builders::{create_inscription_type_0, DaTxs};
 use bitcoin_da::service::{BitcoinService, BitcoinServiceConfig};
 use bitcoin_da::spec::RollupParams;
-use bitcoincore_rpc::{Client, RpcApi};
 use citrea_e2e::config::BitcoinConfig;
 use citrea_e2e::node::NodeKind;
 use citrea_primitives::REVEAL_TX_PREFIX;
@@ -20,9 +16,9 @@ pub(super) enum DaServiceKeyKind {
     Other(String),
 }
 
-const SEQUENCER_DA_PUBLIC_KEY: &str =
+pub const SEQUENCER_DA_PUBLIC_KEY: &str =
     "E9873D79C6D87DC0FB6A5778633389F4453213303DA61F20BD67FC233AA33262";
-const PROVER_DA_PUBLIC_KEY: &str =
+pub(super) const PROVER_DA_PUBLIC_KEY: &str =
     "56D08C2DDE7F412F80EC99A0A328F76688C904BD4D1435281EFC9270EC8C8707";
 
 pub(super) async fn spawn_bitcoin_da_service(
@@ -69,45 +65,4 @@ pub(super) async fn spawn_bitcoin_da_service(
         .spawn_with_graceful_shutdown_signal(|tk| bitcoin_da_service.clone().run_da_queue(rx, tk));
 
     bitcoin_da_service
-}
-
-pub async fn create_complete_tx_with_prefix(
-    client: &Client,
-    body: Vec<u8>,
-    prefix: &[u8],
-) -> anyhow::Result<(Vec<u8>, Vec<u8>)> {
-    let da_private_key = SecretKey::from_str(PROVER_DA_PUBLIC_KEY).unwrap();
-    let change_address = client.get_new_address(None, None).await?.assume_checked();
-    let utxos = client
-        .list_unspent(None, None, None, None, None)
-        .await?
-        .into_iter()
-        .map(Into::into)
-        .collect();
-
-    let result = create_inscription_type_0(
-        body,
-        &da_private_key,
-        None,
-        utxos,
-        change_address,
-        1,
-        1,
-        bitcoin::Network::Regtest,
-        prefix,
-    )?;
-
-    match result {
-        DaTxs::Complete { commit, reveal } => {
-            let signed_raw_commit_tx = client
-                .sign_raw_transaction_with_wallet(&commit, None, None)
-                .await?;
-
-            Ok((
-                signed_raw_commit_tx.hex,
-                bitcoin::consensus::encode::serialize(&reveal.tx),
-            ))
-        }
-        _ => unreachable!("Unexpected result type"),
-    }
 }
